+++ proknow/Patients/Patients.py
@@ -23,7 +23,7 @@ class Patients(object):
         self._requestor = requestor
 
     def _query(self, workspace, query):
-        res, data = self._requestor.get('/workspaces/' + workspace.id + '/patients', query=query)
+        res, data = self._requestor.get('/workspaces/' + workspace.id + '/patients', params=query)
         if res.headers['proknow-has-more'] == 'true': # pragma: no cover (difficult to test w/o lg num of patients)
             next_query = dict(query)
             next_query["next"] = res.headers['proknow-next']
@@ -77,7 +77,7 @@ class Patients(object):
             "sex": sex,
         }
 
-        _, patient = self._requestor.post('/workspaces/' + item.id + '/patients', body=body)
+        _, patient = self._requestor.post('/workspaces/' + item.id + '/patients', json=body)
         return PatientItem(self, item.id, patient)
 
     def delete(self, workspace_id, patient_id):
@@ -169,7 +169,7 @@ class Patients(object):
         assert isinstance(workspace, six.string_types), "`workspace` is required as a string."
 
         item = self._proknow.workspaces.resolve(workspace)
-        _, patients = self._requestor.post('/workspaces/' + item.id + '/patients/lookup', body=mrns)
+        _, patients = self._requestor.post('/workspaces/' + item.id + '/patients/lookup', json=mrns)
         return [PatientSummary(self, item.id, patient) for patient in patients]
 
     def get(self, workspace_id, patient_id):
@@ -263,6 +263,7 @@ class PatientSummary(object):
             patient (dict): A dictionary of patient attributes.
         """
         self._patients = patients
+        self._proknow = self._patients._proknow
         self._requestor = self._patients._requestor
         self._workspace_id = workspace_id
         self._id = patient["id"]
@@ -316,6 +317,45 @@ class PatientSummary(object):
                 patients = [patient.get() for patient in pk.patients.query("Clinical")]
         """
         return self._patients.get(self._workspace_id, self._id)
+
+    def upload(self, path_or_paths, **kwargs):
+        """Initiates an upload or series of uploads to the API for a patient.
+
+        Parameters:
+            path_or_paths (str or list): A path or list of paths such that each path is a directory
+                of files to upload or a path to a file to upload.
+            **kwargs: Keyword arguments to be passed along to the
+                :meth:`proknow.Uploads.Uploads.upload` method.
+
+        Returns:
+            :class:`proknow.Uploads.UploadBatch`: An object used to manage a batch of uploads.
+
+        Raises:
+            AssertionError: If the input parameters are invalid.
+            :class:`proknow.Exceptions.HttpError`: If the HTTP request generated an error.
+            :class:`proknow.Exceptions.InvalidPathError`: If the provided file path is invalid.
+            :class:`proknow.Exceptions.WorkspaceLookupError`: If the workspace with the given
+                name or id could not be found.
+            :class:`proknow.Exceptions.TimeoutExceededError`: If the timeout was exceeded while
+                waiting for the uploads to complete.
+
+        Example:
+            This examples show how to upload a directory of files to a patient::
+
+                from proknow import ProKnow
+                pk = ProKnow('https://example.proknow.com', credentials_file="./credentials.json")
+                patient = pk.patients.lookup('Clinical', ['HNC-0522c0009'])[0]
+                patient.upload("./DICOM")
+        """
+        kwargs["overrides"] = {
+            "patient": {
+                "mrn": self._mrn,
+                "name": self._name,
+                "birth_date": self._birth_date,
+                "sex": self._sex,
+            },
+        }
+        return self._proknow.uploads.upload(self._workspace_id, path_or_paths, **kwargs)
 
 class PatientItem(object):
     """
@@ -407,7 +447,7 @@ class PatientItem(object):
             "sex": self.sex,
             "metadata": self.metadata
         }
-        _, patient = self._requestor.put('/workspaces/' + self._workspace_id + '/patients/' + self._id, body=body)
+        _, patient = self._requestor.put('/workspaces/' + self._workspace_id + '/patients/' + self._id, json=body)
         self._data = patient
         self.mrn = patient["mrn"]
         self.name = patient["name"]
@@ -557,3 +597,43 @@ class PatientItem(object):
             metric = self._proknow.custom_metrics.resolve(key)
             encoded[metric.id] = metadata[key]
         self.metadata = encoded
+
+    def upload(self, path_or_paths, **kwargs):
+        """Initiates an upload or series of uploads to the API for a patient.
+
+        Parameters:
+            path_or_paths (str or list): A path or list of paths such that each path is a directory
+                of files to upload or a path to a file to upload.
+            **kwargs: Keyword arguments to be passed along to the
+                :meth:`proknow.Uploads.Uploads.upload` method.
+
+        Returns:
+            :class:`proknow.Uploads.UploadBatch`: An object used to manage a batch of uploads.
+
+        Raises:
+            AssertionError: If the input parameters are invalid.
+            :class:`proknow.Exceptions.HttpError`: If the HTTP request generated an error.
+            :class:`proknow.Exceptions.InvalidPathError`: If the provided file path is invalid.
+            :class:`proknow.Exceptions.WorkspaceLookupError`: If the workspace with the given
+                name or id could not be found.
+            :class:`proknow.Exceptions.TimeoutExceededError`: If the timeout was exceeded while
+                waiting for the uploads to complete.
+
+        Example:
+            This examples show how to upload a directory of files to a patient::
+
+                from proknow import ProKnow
+                pk = ProKnow('https://example.proknow.com', credentials_file="./credentials.json")
+                patients = pk.patients.lookup('Clinical', ['HNC-0522c0009'])
+                patient = patients[0].get()
+                patient.upload("./DICOM")
+        """
+        kwargs["overrides"] = {
+            "patient": {
+                "mrn": self.mrn,
+                "name": self.name,
+                "birth_date": self.birth_date,
+                "sex": self.sex,
+            },
+        }
+        return self._proknow.uploads.upload(self._workspace_id, path_or_paths, **kwargs)