+++ proknow/Requestor.py
@@ -36,14 +36,12 @@ class Requestor(object):
     def get_base_url(self):
         return self._base_url
 
-    def get(self, route, query=None, headers=None):
+    def get(self, route, **kwargs):
         """Issues an HTTP ``GET`` request.
 
         Parameters:
             route (str): The API route to use in the request.
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            headers (dict, optional): An optional dictionary of request headers to use in the
+            **kwargs (dict, optional): Additional keyword arguments to pass through in the
                 request.
 
         Returns:
@@ -53,19 +51,15 @@ class Requestor(object):
             2. msg (str, dict): the text response or, if the response was JSON, the decoded JSON
                dictionary.
         """
-        r = requests.get(self._base_url + route, params=query, headers=headers, auth=(self._username, self._password))
+        r = requests.get(self._base_url + route, auth=(self._username, self._password), **kwargs)
         return self._handle_response(r)
 
-    def delete(self, route, query=None, body=None, headers=None):
+    def delete(self, route, **kwargs):
         """Issues an HTTP ``DELETE`` request.
 
         Parameters:
             route (str): The API route to use in the request.
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            body (dict, optional): An optional dictionary to be used as a JSON request body for the
-                request.
-            headers (dict, optional): An optional dictionary of request headers to use in the
+            **kwargs (dict, optional): Additional keyword arguments to pass through in the
                 request.
 
         Returns:
@@ -75,19 +69,15 @@ class Requestor(object):
             2. msg (str, dict): the text response or, if the response was JSON, the decoded JSON
                dictionary.
         """
-        r = requests.delete(self._base_url + route, params=query, headers=headers, auth=(self._username, self._password), json=body)
+        r = requests.delete(self._base_url + route, auth=(self._username, self._password), **kwargs)
         return self._handle_response(r)
 
-    def patch(self, route, query=None, body=None, headers=None):
+    def patch(self, route, **kwargs):
         """Issues an HTTP ``PATCH`` request.
 
         Parameters:
             route (str): The API route to use in the request.
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            body (dict, optional): An optional dictionary to be used as a JSON request body for the
-                request.
-            headers (dict, optional): An optional dictionary of request headers to use in the
+            **kwargs (dict, optional): Additional keyword arguments to pass through in the
                 request.
 
         Returns:
@@ -97,19 +87,15 @@ class Requestor(object):
             2. msg (str, dict): the text response or, if the response was JSON, the decoded JSON
                dictionary.
         """
-        r = requests.patch(self._base_url + route, params=query, headers=headers, auth=(self._username, self._password), json=body)
+        r = requests.patch(self._base_url + route, auth=(self._username, self._password), **kwargs)
         return self._handle_response(r)
 
-    def post(self, route, query=None, body=None, headers=None):
+    def post(self, route, **kwargs):
         """Issues an HTTP ``POST`` request.
 
         Parameters:
             route (str): The API route to use in the request.
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            body (dict, optional): An optional dictionary to be used as a JSON request body for the
-                request.
-            headers (dict, optional): An optional dictionary of request headers to use in the
+            **kwargs (dict, optional): Additional keyword arguments to pass through in the
                 request.
 
         Returns:
@@ -119,19 +105,15 @@ class Requestor(object):
             2. msg (str, dict): the text response or, if the response was JSON, the decoded JSON
                dictionary.
         """
-        r = requests.post(self._base_url + route, params=query, headers=headers, auth=(self._username, self._password), json=body)
+        r = requests.post(self._base_url + route, auth=(self._username, self._password), **kwargs)
         return self._handle_response(r)
 
-    def put(self, route, query=None, body=None, headers=None):
+    def put(self, route, **kwargs):
         """Issues an HTTP ``PUT`` request.
 
         Parameters:
             route (str): The API route to use in the request.
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            body (dict, optional): An optional dictionary to be used as a JSON request body for the
-                request.
-            headers (dict, optional): An optional dictionary of request headers to use in the
+            **kwargs (dict, optional): Additional keyword arguments to pass through in the
                 request.
 
         Returns:
@@ -141,7 +123,7 @@ class Requestor(object):
             2. msg (str, dict): the text response or, if the response was JSON, the decoded JSON
                dictionary.
         """
-        r = requests.put(self._base_url + route, params=query, headers=headers, auth=(self._username, self._password), json=body)
+        r = requests.put(self._base_url + route, auth=(self._username, self._password), **kwargs)
         return self._handle_response(r)
 
     def stream(self, route, path):
@@ -150,10 +132,6 @@ class Requestor(object):
         Parameters:
             route (str): The API route to use in the request.
             path (str): The file path to stream the request response
-            query (dict, optional): An optional dictionary of query parameters to use in the
-                request.
-            body (dict, optional): An optional dictionary to be used as a JSON request body for the
-                request.
         """
         with open(path, 'wb') as file:
             with requests.get(self._base_url + route, auth=(self._username, self._password), stream=True) as r: