+++ docs/changelog.rst
@@ -1,6 +1,22 @@
 Release History
 ===============
 
+v0.8.0
+------
+
+**Disclaimer**
+
+All releases in the v0.x.x series are subject to breaking changes from one version to another. After the release of v1.0.0, this project will be subject to `semantic versioning <http://semver.org/>`_.
+
+**New Features and Enhancements**
+
+* Added support for deleting entities
+* Added support for interacting with scorecard templates
+* Added support for getting the current user session
+* Added support for uploading files directly to a patient
+* Reimplemented upload mechanism to utilize the concurrent.futures (`futures` for Python 2) instead of `requests-futures` module
+* Updated Roles for new ProKnow DS version changes (v1.10.0).
+
 v0.7.0
 ------